requirements.txt
# Core Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0

# NLP & Text Processing (use pre-compiled versions)
transformers>=4.20.0
torch>=1.12.0 --index-url https://download.pytorch.org/whl/cpu
tokenizers>=0.13.0
nltk>=3.7.0
spacy>=3.4.0 --no-build-isolation

# Web Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0
urllib3>=1.26.0

# Dashboard
streamlit>=1.12.0

# ML & MLOps
xgboost>=1.6.0
mlflow>=1.30.0
shap>=0.41.0; python_version < "3.12"  # SHAP might not support Python 3.13 yet
joblib>=1.1.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
tqdm>=4.64.0
Jinja2>=3.0.0

# Development (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0